06_ai_agent_collections/03_llama-index/dynamic-tasks-distribution-across-agents/requirements.txt
llama-index
llama-index-llms-openai>=0.3.0
openai
pydantic
beautifulsoup4
google-search-results
python-telegram-bot